DRR - Cpptraj: Replace old DSSP command with new one.
